MOVEONLY: Separate CKeyStore interface from its basic implementation
