+++ src/crypter.h
@@ -6,7 +6,7 @@
 #define BITCOIN_CRYPTER_H
 
 #include "allocators.h"
-#include "keystore.h"
+#include "basic_keystore.h"
 #include "serialize.h"
 
 class uint256;
