+++ src/test/multisig_tests.cpp
@@ -2,8 +2,8 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "basic_keystore.h"
 #include "key.h"
-#include "keystore.h"
 #include "main.h"
 #include "script/script.h"
 #include "script/script_error.h"
