+++ src/test/script_tests.cpp
@@ -5,9 +5,9 @@
 #include "data/script_invalid.json.h"
 #include "data/script_valid.json.h"
 
+#include "basic_keystore.h"
 #include "core_io.h"
 #include "key.h"
-#include "keystore.h"
 #include "main.h"
 #include "script/script.h"
 #include "script/script_error.h"
