+++ src/test/transaction_tests.cpp
@@ -5,9 +5,9 @@
 #include "data/tx_invalid.json.h"
 #include "data/tx_valid.json.h"
 
+#include "basic_keystore.h"
 #include "clientversion.h"
 #include "key.h"
-#include "keystore.h"
 #include "main.h"
 #include "script/script.h"
 #include "script/script_error.h"
